# api/v1/api.gen.go
//go:build go1.22

// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.0 DO NOT EDIT.
package v1

import (
	"fmt"
	"net/http"
	"time"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"
)

// Defines values for ErrorCode.
const (
	ParameterInvalid      ErrorCode = "parameter_invalid"
	ParameterMissing      ErrorCode = "parameter_missing"
	ProcessingError       ErrorCode = "processing_error"
	ResourceAlreadyExists ErrorCode = "resource_already_exists"
	ResourceMissing       ErrorCode = "resource_missing"
)

// Defines values for ErrorType.
const (
	ApiError            ErrorType = "api_error"
	InvalidRequestError ErrorType = "invalid_request_error"
)

// Bookmark Bookmark is a resource that represents a saved URL.
type Bookmark struct {
	CreateTime *time.Time `json:"createTime,omitempty"`

	// Id System-generated unique identifier for the bookmark.
	Id uuid.UUID `json:"id"`

	// ParentId System-generated unique identifier for the parent collection.
	ParentId   *uuid.UUID `json:"parentId,omitempty"`
	Tags       []string   `json:"tags"`
	Title      string     `json:"title"`
	UpdateTime *time.Time `json:"updateTime,omitempty"`
	Url        string     `json:"url"`
}

// BookmarkCreate Resource create operation model.
type BookmarkCreate struct {
	Tags  []string `json:"tags"`
	Title string   `json:"title"`
	Url   string   `json:"url"`
}

// BookmarkUpdate Resource create or update operation model.
type BookmarkUpdate struct {
	Tags  *[]string `json:"tags,omitempty"`
	Title *string   `json:"title,omitempty"`
}

// Collection defines model for Collection.
type Collection struct {
	CreateTime *time.Time `json:"createTime,omitempty"`

	// Id System-generated unique identifier for the collection.
	Id   *uuid.UUID `json:"id,omitempty"`
	Name string     `json:"name"`

	// ParentId System-generated unique identifier for the parent collection.
	ParentId   *uuid.UUID `json:"parentId,omitempty"`
	UpdateTime *time.Time `json:"updateTime,omitempty"`
}

// CollectionChildren defines model for CollectionChildren.
type CollectionChildren struct {
	Bookmarks   []Bookmark   `json:"bookmarks"`
	Collections []Collection `json:"collections"`
}

// CollectionCreate Resource create operation model.
type CollectionCreate struct {
	Name string `json:"name"`

	// ParentId System-generated unique identifier for the parent collection.
	ParentId *uuid.UUID `json:"parentId,omitempty"`
}

// CollectionUpdate Resource create or update operation model.
type CollectionUpdate struct {
	Name *string `json:"name,omitempty"`

	// ParentId System-generated unique identifier for the parent collection.
	ParentId *uuid.UUID `json:"parentId,omitempty"`
}

// Error Error is the response model when an API call is unsuccessful.
type Error struct {
	Code    ErrorCode              `json:"code"`
	Details map[string]interface{} `json:"details"`
	Message string                 `json:"message"`
	Type    ErrorType              `json:"type"`
}

// ErrorCode defines model for ErrorCode.
type ErrorCode string

// ErrorType defines model for ErrorType.
type ErrorType string

// BookmarkKeyBookmarkId defines model for BookmarkKey.bookmarkId.
type BookmarkKeyBookmarkId = uuid.UUID

// BookmarkKeyCollectionId defines model for BookmarkKey.collectionId.
type BookmarkKeyCollectionId = uuid.UUID

// BookmarkKeyUserId defines model for BookmarkKey.userId.
type BookmarkKeyUserId = uuid.UUID

// BookmarkParentKeyCollectionId defines model for BookmarkParentKey.collectionId.
type BookmarkParentKeyCollectionId = uuid.UUID

// BookmarkParentKeyUserId defines model for BookmarkParentKey.userId.
type BookmarkParentKeyUserId = uuid.UUID

// CollectionKeyCollectionId defines model for CollectionKey.collectionId.
type CollectionKeyCollectionId = uuid.UUID

// CollectionKeyUserId defines model for CollectionKey.userId.
type CollectionKeyUserId = uuid.UUID

// CollectionParentKey defines model for CollectionParentKey.
type CollectionParentKey = uuid.UUID

// CollectionsCreateJSONRequestBody defines body for CollectionsCreate for application/json ContentType.
type CollectionsCreateJSONRequestBody = CollectionCreate

// CollectionsUpdateJSONRequestBody defines body for CollectionsUpdate for application/json ContentType.
type CollectionsUpdateJSONRequestBody = CollectionUpdate

// BookmarksCreateJSONRequestBody defines body for BookmarksCreate for application/json ContentType.
type BookmarksCreateJSONRequestBody = BookmarkCreate

// BookmarksUpdateJSONRequestBody defines body for BookmarksUpdate for application/json ContentType.
type BookmarksUpdateJSONRequestBody = BookmarkUpdate

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /users/{userId}/collections)
	CollectionsCreate(w http.ResponseWriter, r *http.Request, userId CollectionParentKey)

	// (GET /users/{userId}/collections/{collectionId})
	CollectionsRead(w http.ResponseWriter, r *http.Request, userId CollectionKeyUserId, collectionId CollectionKeyCollectionId)

	// (PATCH /users/{userId}/collections/{collectionId})
	CollectionsUpdate(w http.ResponseWriter, r *http.Request, userId CollectionKeyUserId, collectionId CollectionKeyCollectionId)

	// (POST /users/{userId}/collections/{collectionId}/bookmarks)
	BookmarksCreate(w http.ResponseWriter, r *http.Request, userId BookmarkParentKeyUserId, collectionId BookmarkParentKeyCollectionId)

	// (GET /users/{userId}/collections/{collectionId}/bookmarks/{bookmarkId})
	BookmarksRead(w http.ResponseWriter, r *http.Request, userId BookmarkKeyUserId, collectionId BookmarkKeyCollectionId, bookmarkId BookmarkKeyBookmarkId)

	// (PATCH /users/{userId}/collections/{collectionId}/bookmarks/{bookmarkId})
	BookmarksUpdate(w http.ResponseWriter, r *http.Request, userId BookmarkKeyUserId, collectionId BookmarkKeyCollectionId, bookmarkId BookmarkKeyBookmarkId)

	// (GET /users/{userId}/collections/{collectionId}/children)
	CollectionsListChildren(w http.ResponseWriter, r *http.Request, userId CollectionKeyUserId, collectionId CollectionKeyCollectionId)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CollectionsCreate operation middleware
func (siw *ServerInterfaceWrapper) CollectionsCreate(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId CollectionParentKey

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CollectionsCreate(w, r, userId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CollectionsRead operation middleware
func (siw *ServerInterfaceWrapper) CollectionsRead(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId CollectionKeyUserId

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// ------------- Path parameter "collectionId" -------------
	var collectionId CollectionKeyCollectionId

	err = runtime.BindStyledParameterWithOptions("simple", "collectionId", r.PathValue("collectionId"), &collectionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collectionId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CollectionsRead(w, r, userId, collectionId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CollectionsUpdate operation middleware
func (siw *ServerInterfaceWrapper) CollectionsUpdate(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId CollectionKeyUserId

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// ------------- Path parameter "collectionId" -------------
	var collectionId CollectionKeyCollectionId

	err = runtime.BindStyledParameterWithOptions("simple", "collectionId", r.PathValue("collectionId"), &collectionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collectionId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CollectionsUpdate(w, r, userId, collectionId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BookmarksCreate operation middleware
func (siw *ServerInterfaceWrapper) BookmarksCreate(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId BookmarkParentKeyUserId

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// ------------- Path parameter "collectionId" -------------
	var collectionId BookmarkParentKeyCollectionId

	err = runtime.BindStyledParameterWithOptions("simple", "collectionId", r.PathValue("collectionId"), &collectionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collectionId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BookmarksCreate(w, r, userId, collectionId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BookmarksRead operation middleware
func (siw *ServerInterfaceWrapper) BookmarksRead(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId BookmarkKeyUserId

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// ------------- Path parameter "collectionId" -------------
	var collectionId BookmarkKeyCollectionId

	err = runtime.BindStyledParameterWithOptions("simple", "collectionId", r.PathValue("collectionId"), &collectionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collectionId", Err: err})
		return
	}

	// ------------- Path parameter "bookmarkId" -------------
	var bookmarkId BookmarkKeyBookmarkId

	err = runtime.BindStyledParameterWithOptions("simple", "bookmarkId", r.PathValue("bookmarkId"), &bookmarkId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bookmarkId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BookmarksRead(w, r, userId, collectionId, bookmarkId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BookmarksUpdate operation middleware
func (siw *ServerInterfaceWrapper) BookmarksUpdate(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId BookmarkKeyUserId

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// ------------- Path parameter "collectionId" -------------
	var collectionId BookmarkKeyCollectionId

	err = runtime.BindStyledParameterWithOptions("simple", "collectionId", r.PathValue("collectionId"), &collectionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collectionId", Err: err})
		return
	}

	// ------------- Path parameter "bookmarkId" -------------
	var bookmarkId BookmarkKeyBookmarkId

	err = runtime.BindStyledParameterWithOptions("simple", "bookmarkId", r.PathValue("bookmarkId"), &bookmarkId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bookmarkId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BookmarksUpdate(w, r, userId, collectionId, bookmarkId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CollectionsListChildren operation middleware
func (siw *ServerInterfaceWrapper) CollectionsListChildren(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userId CollectionKeyUserId

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// ------------- Path parameter "collectionId" -------------
	var collectionId CollectionKeyCollectionId

	err = runtime.BindStyledParameterWithOptions("simple", "collectionId", r.PathValue("collectionId"), &collectionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "collectionId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CollectionsListChildren(w, r, userId, collectionId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/users/{userId}/collections", wrapper.CollectionsCreate)
	m.HandleFunc("GET "+options.BaseURL+"/users/{userId}/collections/{collectionId}", wrapper.CollectionsRead)
	m.HandleFunc("PATCH "+options.BaseURL+"/users/{userId}/collections/{collectionId}", wrapper.CollectionsUpdate)
	m.HandleFunc("POST "+options.BaseURL+"/users/{userId}/collections/{collectionId}/bookmarks", wrapper.BookmarksCreate)
	m.HandleFunc("GET "+options.BaseURL+"/users/{userId}/collections/{collectionId}/bookmarks/{bookmarkId}", wrapper.BookmarksRead)
	m.HandleFunc("PATCH "+options.BaseURL+"/users/{userId}/collections/{collectionId}/bookmarks/{bookmarkId}", wrapper.BookmarksUpdate)
	m.HandleFunc("GET "+options.BaseURL+"/users/{userId}/collections/{collectionId}/children", wrapper.CollectionsListChildren)

	return m
}